Address Sabir's feedback
