+++ spark/src/main/scala/org/apache/spark/sql/delta/PreDowngradeTableFeatureCommand.scala
@@ -217,6 +217,7 @@ case class TypeWideningPreDowngradeCommand(table: DeltaTableV2)
 case class ColumnMappingUsageTrackingPreDowngradeCommand(table: DeltaTableV2)
     extends PreDowngradeTableFeatureCommand {
   override def removeFeatureTracesIfNeeded(): Boolean = {
+    if (ColumnMappingUsageTrackingTableFeature.validateRemoval(table.initialSnapshot)) return false
     // Drop the internal table property to ensure it's no longer there if the feature ever gets
     // re-enabled. Otherwise we might incorrectly think that no columns were dropped or removed if
     // the feature is disabled, a column is dropped, and then the feature is re-enabled again.
